move calculateAmountWithTransferFees to a function
