+++ packages/distributor/solana/utils.ts
@@ -1,11 +1,13 @@
+import { TransferFeeConfig } from "@solana/spl-token";
 import { SignerWalletAdapter } from "@solana/wallet-adapter-base";
 import { Connection, Keypair, PublicKey, Transaction, VersionedTransaction } from "@solana/web3.js";
 import { ContractError } from "@streamflow/common";
 import { ConfirmationParams, signAndExecuteTransaction, ThrottleParams } from "@streamflow/common/solana";
 
 import { fromTxError } from "./generated/errors";
+import { ONE_IN_BASIS_POINTS } from "./constants";
 
-export const ceilN = (n: bigint, d: bigint): bigint => n / d + (n % d ? BigInt(1) : BigInt(0));
+export const divCeilN = (n: bigint, d: bigint): bigint => n / d + (n % d ? BigInt(1) : BigInt(0));
 
 export function getDistributorPda(programId: PublicKey, mint: PublicKey, version: number): PublicKey {
   // Constructing the seed for the PDA
@@ -53,4 +55,29 @@ export async function wrappedSignAndExecuteTransaction(
     }
     throw err;
   }
+}
+
+export async function calculateAmountWithTransferFees(
+  connection: Connection,
+  transferFeeConfig: TransferFeeConfig,
+  transferAmount: bigint,
+): Promise<{ transferAmount: bigint; feeCharged: bigint }> {
+  const epoch = await connection.getEpochInfo();
+  const transferFee =
+    epoch.epoch >= transferFeeConfig.newerTransferFee.epoch
+      ? transferFeeConfig.newerTransferFee
+      : transferFeeConfig.olderTransferFee;
+  const transferFeeBasisPoints = BigInt(transferFee.transferFeeBasisPoints);
+  let feeCharged = BigInt(0);
+
+  if (transferFeeBasisPoints !== BigInt(0)) {
+    const numerator = transferAmount * ONE_IN_BASIS_POINTS;
+    const denominator = ONE_IN_BASIS_POINTS - transferFeeBasisPoints;
+    const rawPreFeeAmount = divCeilN(numerator, denominator);
+    const fee = rawPreFeeAmount - transferAmount;
+    transferAmount = rawPreFeeAmount;
+    feeCharged = fee > transferFee.maximumFee ? transferFee.maximumFee : fee;
+  }
+
+  return { transferAmount, feeCharged };
 }
